pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sae-scoping"
version = "2020.0.0"
dependencies = [
  "beartype==0.14.1",
  "click==8.2.1",
  "datasets==4.0.0",
  "ipython==9.4.0",
  "jaxtyping==0.3.2",
  "litellm==1.74.7",
  "notebook==7.4.4",
  "numpy==1.26.4",
  "orjson==3.11.0",
  "pandas==2.3.1",
  "peft==0.16.0",
  "plotly==6.3.0",
  "pydantic==2.12.4",
  "pytest==8.4.1",
  "requests==2.32.4",
  "ruff==0.14.2",
  "sae-lens==6.22.3",
  "safetensors==0.5.3",
  "scikit-learn==1.7.1",
  "seaborn==0.13.2",
  "torch==2.7.1",
  "transformers==4.56.1",
  "trl==0.22.2",
  "wandb==0.21.0",
  "eai-sparsify==1.3.0",
  "pyyaml>=6.0",
]
requires-python = ">=3.12"
authors = [
  {name = "Adriano Hernandez", email = "adrianoh@mit.edu"},
]
maintainers = [
  {name = "Adriano Hernandez", email = "adrianoh@mit.edu"}
]
description = "A minimal toolkit for scoping language models using Sparse Autoencoders"
readme = "README.md"
license = "GPL-3.0"
license-files = ["LICENSE.md"]
keywords = ["SAE", "Scoping", "Language Models"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.12"
]
[project.urls]
Homepage = "https://github.com/4gatepylon/SAEScoping"